Did not lock sections but did refactoring
