+++ app/(2 Authenticated)/learn/page.tsx
@@ -1,467 +1,19 @@
 "use client"
-import {
-  ActiveBookSvg,
-  ActiveDumbbellSvg,
-  ActiveTreasureSvg,
-  ActiveTrophySvg,
-  CheckmarkSvg,
-  FastForwardSvg,
-  GoldenBookSvg,
-  GoldenDumbbellSvg,
-  GoldenTreasureSvg,
-  GoldenTrophySvg,
-  GuidebookSvg,
-  LessonCompletionSvg0,
-  LessonCompletionSvg1,
-  LessonCompletionSvg2,
-  LessonCompletionSvg3,
-  LockSvg,
-  LockedBookSvg,
-  LockedDumbbellSvg,
-  LockedTreasureSvg,
-  LockedTrophySvg,
-  StarSvg,
-} from "@/components/SVGs"
 import { BottomBar } from "@/components/learn/BottomBar"
 import { LeftBar } from "@/components/learn/LeftBar"
 import { RightBar } from "@/components/learn/RightBar"
 import { TopBar } from "@/components/learn/TopBar"
-import languages from "@/utils/languages"
-import type { Tile, TileType, Unit } from "@/utils/units"
-import { units } from "@/utils/units"
+import UnitSection from "@/components/learn/home/UnitSection"
 import { type NextPage } from "next"
-import Link from "next/link"
-import { useRouter } from "next/navigation"
-import React, {
-  Fragment,
-  useCallback,
-  useEffect,
-  useRef,
-  useState,
-} from "react"
+import React, { useEffect, useState } from "react"
 import { useGetUnitsQuery } from "@/store/slices/UnitSlice"
 import { useGetSectionsQuery } from "@/store/slices/SectionSlice"
-
-type TileStatus = "LOCKED" | "ACTIVE" | "COMPLETE"
-
-const tileStatus = (tile: Tile, lessonsCompleted: number): TileStatus => {
-  const lessonsPerTile = 4
-  const tilesCompleted = Math.floor(lessonsCompleted / lessonsPerTile)
-  const tiles = units.flatMap((unit: any) => unit.tiles)
-  const tileIndex = tiles.findIndex((t) => t === tile)
-
-  if (tileIndex < tilesCompleted) {
-    return "COMPLETE"
-  }
-  if (tileIndex > tilesCompleted) {
-    return "LOCKED"
-  }
-  return "ACTIVE"
-}
-
-const TileIcon = ({
-  tileType,
-  status,
-}: {
-  tileType: TileType
-  status: TileStatus
-  // eslint-disable-next-line no-undef
-}): React.ReactNode => {
-  switch (tileType) {
-    case "star":
-      return status === "COMPLETE" ? (
-        <CheckmarkSvg />
-      ) : status === "ACTIVE" ? (
-        <StarSvg />
-      ) : (
-        <LockSvg />
-      )
-    case "book":
-      return status === "COMPLETE" ? (
-        <GoldenBookSvg />
-      ) : status === "ACTIVE" ? (
-        <ActiveBookSvg />
-      ) : (
-        <LockedBookSvg />
-      )
-    case "dumbbell":
-      return status === "COMPLETE" ? (
-        <GoldenDumbbellSvg />
-      ) : status === "ACTIVE" ? (
-        <ActiveDumbbellSvg />
-      ) : (
-        <LockedDumbbellSvg />
-      )
-    case "fast-forward":
-      return status === "COMPLETE" ? (
-        <CheckmarkSvg />
-      ) : status === "ACTIVE" ? (
-        <StarSvg />
-      ) : (
-        <FastForwardSvg />
-      )
-    case "treasure":
-      return status === "COMPLETE" ? (
-        <GoldenTreasureSvg />
-      ) : status === "ACTIVE" ? (
-        <ActiveTreasureSvg />
-      ) : (
-        <LockedTreasureSvg />
-      )
-    case "trophy":
-      return status === "COMPLETE" ? (
-        <GoldenTrophySvg />
-      ) : status === "ACTIVE" ? (
-        <ActiveTrophySvg />
-      ) : (
-        <LockedTrophySvg />
-      )
-  }
-}
-
-const tileLeftClassNames = [
-  "left-0",
-  "left-[-45px]",
-  "left-[-70px]",
-  "left-[-45px]",
-  "left-0",
-  "left-[45px]",
-  "left-[70px]",
-  "left-[45px]",
-] as const
-
-type TileLeftClassName = (typeof tileLeftClassNames)[number]
-
-const getTileLeftClassName = ({
-  index,
-  unitNumber,
-  tilesLength,
-}: {
-  index: number
-  unitNumber: number
-  tilesLength: number
-}): TileLeftClassName => {
-  if (index >= tilesLength - 1) {
-    return "left-0"
-  }
-
-  const classNames =
-    unitNumber % 2 === 1
-      ? tileLeftClassNames
-      : [...tileLeftClassNames.slice(4), ...tileLeftClassNames.slice(0, 4)]
-
-  return classNames[index % classNames.length] ?? "left-0"
-}
-
-const tileTooltipLeftOffsets = [140, 95, 70, 95, 140, 185, 210, 185] as const
-
-type TileTooltipLeftOffset = (typeof tileTooltipLeftOffsets)[number]
-
-const getTileTooltipLeftOffset = ({
-  index,
-  unitNumber,
-  tilesLength,
-}: {
-  index: number
-  unitNumber: number
-  tilesLength: number
-}): TileTooltipLeftOffset => {
-  if (index >= tilesLength - 1) {
-    return tileTooltipLeftOffsets[0]
-  }
-
-  const offsets =
-    unitNumber % 2 === 1
-      ? tileTooltipLeftOffsets
-      : [
-          ...tileTooltipLeftOffsets.slice(4),
-          ...tileTooltipLeftOffsets.slice(0, 4),
-        ]
-
-  return offsets[index % offsets.length] ?? tileTooltipLeftOffsets[0]
-}
-
-const getTileColors = ({
-  tileType,
-  status,
-  defaultColors,
-}: {
-  tileType: TileType
-  status: TileStatus
-  defaultColors: `border-${string} bg-${string}`
-}): `border-${string} bg-${string}` => {
-  switch (status) {
-    case "LOCKED":
-      if (tileType === "fast-forward") return defaultColors
-      return "border-[#b7b7b7] bg-[#e5e5e5]"
-    case "COMPLETE":
-      return "border-yellow-500 bg-yellow-400"
-    case "ACTIVE":
-      return defaultColors
-  }
-}
-
-const TileTooltip = ({
-  selectedTile,
-  index,
-  section_id,
-  unitNumber,
-  tilesLength,
-  description,
-  status,
-  closeTooltip,
-}: {
-  selectedTile: number | null
-  index: number
-  section_id: number
-  unitNumber: number
-  tilesLength: number
-  description: string
-  status: TileStatus
-  closeTooltip: () => void
-}) => {
-  const tileTooltipRef = useRef<HTMLDivElement | null>(null)
-
-  useEffect(() => {
-    const containsTileTooltip = (event: MouseEvent) => {
-      if (selectedTile !== index) return
-      const clickIsInsideTooltip = tileTooltipRef.current?.contains(
-        event.target as Node,
-      )
-      if (clickIsInsideTooltip) return
-      closeTooltip()
-    }
-
-    window.addEventListener("click", containsTileTooltip, true)
-    return () => window.removeEventListener("click", containsTileTooltip, true)
-  }, [selectedTile, tileTooltipRef, closeTooltip, index])
-
-  const unit = units.find((unit) => unit.unitNumber === unitNumber)
-  const activeBackgroundColor = unit?.backgroundColor ?? "bg-green-500"
-  const activeTextColor = unit?.textColor ?? "text-green-500"
-
-  return (
-    <div
-      className={[
-        "relative h-0 w-full",
-        index === selectedTile ? "" : "invisible",
-      ].join(" ")}
-      ref={tileTooltipRef}>
-      <div
-        className={[
-          "absolute z-30 flex w-[300px] flex-col gap-4 rounded-xl p-4 font-bold transition-all duration-300",
-          status === "ACTIVE"
-            ? activeBackgroundColor
-            : status === "LOCKED"
-            ? "border-2 border-gray-200 bg-gray-100"
-            : "bg-yellow-400",
-          index === selectedTile ? "top-4 scale-100" : "-top-14 scale-0",
-        ].join(" ")}
-        style={{ left: "calc(50% - 150px)" }}>
-        <div
-          className={[
-            "absolute top-[-8px] left-[140px] h-4 w-4 rotate-45",
-            status === "ACTIVE"
-              ? activeBackgroundColor
-              : status === "LOCKED"
-              ? "border-t-2 border-l-2 border-gray-200 bg-gray-100"
-              : "bg-yellow-400",
-          ].join(" ")}
-          style={{
-            left: getTileTooltipLeftOffset({ index, unitNumber, tilesLength }),
-          }}></div>
-        <div
-          className={[
-            "text-lg",
-            status === "ACTIVE"
-              ? "text-white"
-              : status === "LOCKED"
-              ? "text-gray-400"
-              : "text-yellow-600",
-          ].join(" ")}>
-          {description}
-        </div>
-        {status === "ACTIVE" ? (
-          <Link
-            href='/learn/quiz'
-            className={[
-              "flex w-full items-center justify-center rounded-xl border-b-4 border-gray-200 bg-white p-3 uppercase",
-              activeTextColor,
-            ].join(" ")}>
-            Start +10 XP
-          </Link>
-        ) : status === "LOCKED" ? (
-          <button
-            className='w-full rounded-xl bg-gray-200 p-3 uppercase text-gray-400'
-            disabled>
-            Locked
-          </button>
-        ) : (
-          <Link
-            href={`/learn/section/${section_id}`}
-            className='flex w-full items-center justify-center rounded-xl border-b-4 border-yellow-200 bg-white p-3 uppercase text-yellow-400'>
-            Practice +5 XP
-          </Link>
-        )}
-      </div>
-    </div>
-  )
-}
-
-const UnitSection = ({ unit }: { unit: Unit }): React.ReactNode => {
-  const router = useRouter()
-
-  const [selectedTile, setSelectedTile] = useState<null | number>(null)
-
-  useEffect(() => {
-    const unselectTile = () => setSelectedTile(null)
-    window.addEventListener("scroll", unselectTile)
-    return () => window.removeEventListener("scroll", unselectTile)
-  }, [])
-
-  const closeTooltip = useCallback(() => setSelectedTile(null), [])
-
-  const [lessonsCompleted] = useState(6)
-
-  const increaseLessonsCompleted = 7
-  const increaseLingots = 8
-  return (
-    <>
-      <UnitHeader
-        unitNumber={unit.unitNumber}
-        description={unit.description}
-        backgroundColor={unit.backgroundColor}
-        borderColor={unit.borderColor}
-      />
-      <div className='relative mt-[20px] mb-8 flex max-w-2xl flex-col items-center gap-4'>
-        {unit.sections.map((tile: any, i: any): React.ReactNode => {
-          const status = tileStatus(tile, lessonsCompleted)
-          return (
-            <Fragment key={i}>
-              {(() => {
-                switch (tile.type) {
-                  case "star":
-                  case "book":
-                  case "dumbbell":
-                  case "trophy":
-                  case "fast-forward":
-                    if (tile.type === "trophy" && status === "COMPLETE") {
-                      return (
-                        <div className='relative'>
-                          <TileIcon tileType={tile.type} status={status} />
-                          <div className='absolute top-6 left-0 right-0 flex justify-center text-lg font-bold text-yellow-700'>
-                            {unit.unitNumber}
-                          </div>
-                        </div>
-                      )
-                    }
-                    return (
-                      <div
-                        className={[
-                          "relative -mb-4 h-[93px] w-[98px]",
-                          getTileLeftClassName({
-                            index: i,
-                            unitNumber: unit.unitNumber,
-                            tilesLength: unit.sections.length,
-                          }),
-                        ].join(" ")}>
-                        {tile.type === "fast-forward" && status === "LOCKED" ? (
-                          <HoverLabel
-                            text='Jump here?'
-                            textColor={unit.textColor}
-                          />
-                        ) : selectedTile !== i && status === "ACTIVE" ? (
-                          <HoverLabel text='Start' textColor={unit.textColor} />
-                        ) : null}
-                        <LessonCompletionSvg
-                          lessonsCompleted={lessonsCompleted}
-                          status={status}
-                        />
-                        <button
-                          className={[
-                            "absolute m-3 rounded-full border-b-8 p-4",
-                            getTileColors({
-                              tileType: tile.type,
-                              status,
-                              defaultColors: `${unit.borderColor} ${unit.backgroundColor}`,
-                            }),
-                          ].join(" ")}
-                          onClick={() => {
-                            if (
-                              tile.type === "fast-forward" &&
-                              status === "LOCKED"
-                            ) {
-                              void router.push(
-                                `/lesson?fast-forward=${unit.unitNumber}`,
-                              )
-                              return
-                            }
-                            setSelectedTile(i)
-                          }}>
-                          <TileIcon tileType={tile.type} status={status} />
-                          <span className='sr-only'>Show lesson</span>
-                        </button>
-                      </div>
-                    )
-                  case "treasure":
-                    return (
-                      <div
-                        className={[
-                          "relative -mb-4",
-                          getTileLeftClassName({
-                            index: i,
-                            unitNumber: unit.unitNumber,
-                            tilesLength: unit.sections.length,
-                          }),
-                        ].join(" ")}
-                        role='button'
-                        tabIndex={status === "ACTIVE" ? 0 : undefined}
-                        aria-hidden={status !== "ACTIVE"}
-                        aria-label={
-                          status === "ACTIVE" ? "Collect reward" : ""
-                        }>
-                        {status === "ACTIVE" && (
-                          <HoverLabel text='Open' textColor='text-yellow-400' />
-                        )}
-                        <TileIcon tileType={tile.type} status={status} />
-                      </div>
-                    )
-                }
-              })()}
-              <TileTooltip
-                selectedTile={selectedTile}
-                index={i}
-                section_id={tile.id}
-                unitNumber={unit.unitNumber}
-                tilesLength={unit.sections.length}
-                description={(() => {
-                  switch (tile.type) {
-                    case "book":
-                    case "dumbbell":
-                    case "star":
-                      return tile.title
-                    case "fast-forward":
-                      return status === "LOCKED" ? "Jump here?" : tile.title
-                    case "trophy":
-                      return `Unit ${unit.unitNumber} review`
-                    case "treasure":
-                      return ""
-                  }
-                })()}
-                status={status}
-                closeTooltip={closeTooltip}
-              />
-            </Fragment>
-          )
-        })}
-      </div>
-    </>
-  )
-}
 
 const Learn: NextPage = () => {
   const { data: unitz } = useGetUnitsQuery<any>()
   const { data: sections } = useGetSectionsQuery<any>()
+
+  // TODO: Send "sectionCompletedStatus" to unitSection component, it could be "LOCKED" or "COMPLETE". Show "ACTIVE" only when the user hasn't started anything, only on the first one
 
   const [formattedData, setFormattedData] = useState<any>([])
 
@@ -528,98 +80,3 @@ const Learn: NextPage = () => {
 }
 
 export default Learn
-
-const LessonCompletionSvg = ({
-  lessonsCompleted,
-  status,
-  style = {},
-}: {
-  lessonsCompleted: number
-  status: TileStatus
-  style?: React.HTMLAttributes<SVGElement>["style"]
-}) => {
-  if (status !== "ACTIVE") {
-    return null
-  }
-  switch (lessonsCompleted % 4) {
-    case 0:
-      return <LessonCompletionSvg0 style={style} />
-    case 1:
-      return <LessonCompletionSvg1 style={style} />
-    case 2:
-      return <LessonCompletionSvg2 style={style} />
-    case 3:
-      return <LessonCompletionSvg3 style={style} />
-    default:
-      return null
-  }
-}
-
-const HoverLabel = ({
-  text,
-  textColor,
-}: {
-  text: string
-  textColor: `text-${string}`
-}) => {
-  const hoverElement = useRef<HTMLDivElement | null>(null)
-  const [width, setWidth] = useState(72)
-
-  useEffect(() => {
-    setWidth(hoverElement.current?.clientWidth ?? width)
-  }, [hoverElement.current?.clientWidth, width])
-
-  return (
-    <div
-      className={`absolute z-10 w-max animate-bounce rounded-lg border-2 border-gray-200 bg-white py-2 px-3 font-bold uppercase ${textColor}`}
-      style={{
-        top: "-25%",
-        left: `calc(50% - ${width / 2}px)`,
-      }}
-      ref={hoverElement}>
-      {text}
-      <div
-        className='absolute h-3 w-3 rotate-45 border-b-2 border-r-2 border-gray-200 bg-white'
-        style={{ left: "calc(50% - 8px)", bottom: "-8px" }}></div>
-    </div>
-  )
-}
-
-const UnitHeader = ({
-  unitNumber,
-  description,
-  backgroundColor,
-  borderColor,
-}: {
-  unitNumber: number
-  description: string
-  backgroundColor: `bg-${string}`
-  borderColor: `border-${string}`
-}) => {
-  const language = languages.filter((lang) => lang.code === "en")[0]
-  return (
-    <article
-      className={[
-        "max-w-2xl text-white py-2 sm:rounded-xl",
-        backgroundColor,
-      ].join(" ")}>
-      <header className='flex items-center justify-between gap-4 p-4'>
-        <div className='flex flex-col gap-1'>
-          <h2 className='text-2xl font-bold'>Unit {unitNumber}</h2>
-          <p className='text-lg'>{description}</p>
-        </div>
-        <Link
-          href={`learn/guidebook/${language.code}/${unitNumber}`}
-          className={[
-            "flex items-center gap-3 rounded-2xl border-2 border-b-4 p-3 transition hover:text-gray-100",
-            borderColor,
-          ].join(" ")}>
-          <GuidebookSvg />
-          <span className='sr-only font-bold uppercase lg:not-sr-only'>
-            Guidebook
-          </span>
-        </Link>
-      </header>
-    </article>
-  )
-}
