+++ components/index.ts
@@ -1,4 +1,3 @@
-export { default as Units } from "./learn/Units"
 export { default as SideBar } from "./learn/SideBar"
 export { default as Navbar } from "./navbar/Navbar"
 export { default as ForgetPassword } from "./emails/ForgetPassword"
